+++ CHANGES.md
@@ -9,6 +9,7 @@
 * `#[js_json(default = "None")]` attribute to `AutoJsJson`
 * All http methods in `FetchMethod`
 * `history_replace` method in `Driver`
+* vertigo-cli: `add-watch-path` to `watch` command
 
 ### Fixed